refactor: Модуль curtains.js переименован в change-screens.js
Переименование для расширении функциональности модуля.
